Create user profile component and route

+++ frontend/components/checkin/checkin_index_item.jsx
@@ -11,9 +11,11 @@ class CheckinIndexItem extends React.Component{
 
   handleDelete(e){
     e.preventDefault();
-    this.props.deleteCheckin(this.props.checkin.id).then(() =>{
-      this.props.history.push("/home");
-    });
+    if (confirm("Are you sure you want to delete this checkin?")){
+      this.props.deleteCheckin(this.props.checkin.id).then(() =>{
+        this.props.history.push("/home");
+      });
+    }
   }
 
   calculateRating(rating){
@@ -75,7 +77,7 @@ class CheckinIndexItem extends React.Component{
       let editUrl = `/checkins/${this.props.checkin.id}/edit`;
       viewEditCheckin = <div className="edit-delete-checkin">
         <Link className="edit-checkin" to={editUrl}>Edit Checkin</Link>
-        <h4 onClick={this.handleDelete} className="dcheckin-elete">Delete Checkin</h4>
+        <h4 onClick={this.handleDelete} className="checkin-delete">Delete Checkin</h4>
       </div>;
     }
